server-side validations with zod

+++ src/actions/actions.ts
@@ -4,13 +4,24 @@ import { revalidatePath } from 'next/cache';
 import { sleep } from '@/lib/utils';
 import { PetEssentials } from '@/lib/types';
 import { Pet } from '@prisma/client';
+import { PetFormSchema, PetIdSchema } from '@/lib/validations';
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
+  //unknown is a safer type for a backend endpoint instead of assuming the data will be the correct type
   await sleep(1000)
+
+  const validatedPet = PetFormSchema.safeParse(pet);
+
+  if (!validatedPet.success) {
+    return {
+      message: "Invalid pet data."
+    }
+  }
+
   try {
 
     await prisma.pet.create({
-       data: pet
+       data: validatedPet.data
     })
   } catch (error) {
     return {
@@ -22,14 +33,26 @@ export async function addPet(pet: PetEssentials) {
  revalidatePath('/app', "layout");
 }
 
-export async function editPet(petId: Pet['id'], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000)
+
+  const validatedPetId = PetIdSchema.safeParse(petId);
+  
+  const validatedPet = PetFormSchema.safeParse(newPetData);
+
+  if (!validatedPet.success || !validatedPetId.success) {
+    return {
+      message: "Invalid pet data."
+    }
+  }
+
+
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData
+      data: validatedPet.data
     });
   }
   catch (error) {
@@ -41,12 +64,20 @@ export async function editPet(petId: Pet['id'], newPetData: PetEssentials) {
 
 }
 
-export async function deletePet(petId: Pet['id']) {
+export async function deletePet(petId: unknown) {
   await sleep(1000)
+
+  const validatedPetId = PetIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data."
+    }
+  }
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   }
